Fixed lazyLoading for CustomTab Fixed jsonTab command
