+++ packages/admin/src-admin/package.json
@@ -109,5 +109,5 @@
             }
         ]
     ],
-    "version": "7.6.0"
+    "version": "7.6.1"
 }
